pyproject.toml
[project]
name = "funding-service-design-fund-application-builder"
version = "0.0.1"
description = "Prototype Fund Application Builder (FAB) app for building application forms"
authors = ["MHCLG"]
license = "MIT License"

requires-python = "~=3.10.0"
dependencies = [
    "airium==0.2.6",
    "beautifulsoup4==4.13.3",
    "flask-migrate==4.1.0",
    "flask-sqlalchemy==3.1.1",
    "flask-wtf==1.2.2",
    "flask==3.1.0",
    "flask-talisman==1.1.0",
    "funding-service-design-utils==6.0.2",
    "jsonschema==4.23.0",
    "psycopg2-binary==2.9.10",
    "requests==2.32.3",
    "sentry-sdk==2.22.0",
    "sqlalchemy[mypy]==2.0.38",
    "flask-debugtoolbar==0.16.0",
    #
    # If updating either of the govuk-frontend-* packages below, check whether any changes need to be synced into the
    # govuk_frontend_ext module.
    "govuk-frontend-jinja==3.4.1",
    "govuk-frontend-wtf==3.2.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
    "E", # pycodestyle
    "W", # pycodestyle
    "F", # pyflakes
    "I", # isort
    "B", # flake8-bugbear
    "C90", # mccabe cyclomatic complexity
    "G", # flake8-logging-format
]
ignore = []
exclude = [
    "venv*",
    ".venv*",
    "__pycache__",
]
mccabe.max-complexity = 12

[tool.uv]

[dependency-groups]
dev = [
    "debugpy==1.8.12",
    "djlint==1.36.4",
    "invoke==2.2.0",
    "pre-commit==4.1.0",
    "pytest-env==1.1.5",
    "pytest-html==4.1.1",
    "pytest-mock==3.14.0",
    "pytest==8.3.5",
    "python-dotenv==1.0.1",
    "ruff==0.9.9",
    "pytest-playwright==0.7.0",
    "Faker==36.1.1",
]
